audio: Do not use StreamSwitcher for StreamRemoteSubmix (V2)
Since use of StreamSwitcher causes the worker thread to be changed
during connected device change, its use must be avoided. We intend
to remove StreamSwitcher completely in future.

In the second revision of this change make sure that the remote
submix correctly handles matching of the configurations between
the ends of a pipe and rejects non-matching configurations early.

Bug: 300130515
Bug: 338974476
Bug: 368723297
Bug: 369272078
Bug: 369289912
Bug: 369964381
Bug: 392011259
Bug: 403165213
Bug: 434879819
Bug: 436373649
Bug: 443276469
Test: atest CtsMediaAudioTestCases
Test: atest CtsVirtualDevicesAudioTestCases
Test: atest VtsHalAudioCoreTargetTest
Change-Id: Ifc882c5fccf761adc267a8668631a05aea83eb7f