Use the default destructor for AidlCallbackHandler and add a log in the death handler.
The destructor and the death handler are the
only functions that might be called when the
service is exiting. Both methods currently
attempt to retrieve the static global lock.

In a rare race condition, the static lock may
be deleted before the method is called. To
limit the potential for this issue, we can
avoid retrieving the lock in the destructor
by using the default destructor. The death
handler seems like a less likely culprit, so
we can add a log there for now.

Bug: 384015760
Test: m
Change-Id: I0073a461f98a6c22e6be7e94747a700dfea0753b