Update Vibrator AIDL to improve rust support
Introduce @nullable annotation to AIDL interfaces, to fix the rust
backend generation and allow null parameters to be passed. The framework
will send null values for vibration callbacks when the support is
not available in the HAL implementation. Existing VTS cover this
scenario, passing null when support is not available or the callback is
not important for the test.

Enable gen_mockall for rust backend generator to allow the framework to
use the rust client and mock the service for unit tests.

Fix: 398167653
Test: VtsHalVibratorTargetTest
      VtsHalVibratorManagerTargetTest
Flag: EXEMPT AIDL and build file changes
Change-Id: If78abd16f02f8f2f689541dfcc2bda16ccb5dc0c