Relax requirement for DICE in keymint
It turns out "mid-tier" chips are based on older silicon/ROM that
cannot support DICE. OEMs are pairing these parts with an SE and
enabling StrongBox. This means the old requirement cannot be satisfied.

Instead of requiring DICE when StrongBox is enabled, we look for pVM
support (which already requires DICE).

Bug: 401663739
Bug: 402031644
Bug: 378683581
Test: atest
Change-Id: I5fb7617aff8b219e29eb741f7c520a38ece70532