audio: Update VTS to send zero burst value for MMap correctly
A follow up for ag/I11ab819f9fdf4131537d77ad2c66626ffa29992c.
For input MMap streams the test SendInvalidCommand got broken
because the modification was always resetting the burst value
size to 0, even when the test was intentionally setting it to
an invalid value to test the HAL implementation error handling.

This was not discovered in presubmit testing because the test
is arbitrarily picking up one mix port which in our case was not
MMap. The issue was found on a different HAL configuration.
To prevent this, enhance the SendInvalidCommand to test MMap and
non-MMap specifically.

Flag: EXEMPT bugfix
Bug: 418983038
Test: atest VtsHalAudioCoreTargetTest
Change-Id: I19f1e53947954d2d3dbf3ec9555828bb9b4b1477