Relax battery serial number validity requirements
It's valid for a battery to have a serial number shorter than 6 characters,
so it's wrong to test that it's at least that long. For example, the marasov
device with serial number R8NXCV05W978323 has a battery with serial number
"139C" which is perfectly valid. Reduce the expectation to at least 4
characters, since a very short string doesn't seem like it would be useful
and could still indicate a problem.

Bug: 417324494
Flag: TEST_ONLY
Change-Id: Ie50df424e4b379ad4c5f2ab609514428afe6649a