wifip2p: Fixes in SupplicantP2pNetworkTargetTest
The changes are,
0. Take the last added network in the p2p_supplicant.conf.
This is to ensure that the test is picking the persistent
network added in the setup() function.
1. The isCurrent() check is not relevant for the P2P device
interface as the current network is set on the P2P group
interface. This commit removes the unnecessary check,
relying solely on the API call's success.
2. Added P2pIface#saveConfig() call after calling
P2pIface#removeNetwork() to update the configuration
in p2p_supplicant.conf.

Bug: 422047384
Test: atest VtsHalWifiSupplicantP2pNetworkTargetTest
Change-Id: I0d5c8c192b8c901d6a96536306b00784dc0573d9