Add wakelock in hci_router when handing TxTask
A potential race condition exists where the system can suspend after a
TxTask is posted to the HciRouter but before it's processed by the TX
thread. This suspension can cause an HCI timeout.

By acquiring a wakelock for the duration of the TxTask, we ensure the
system remains active, preventing premature suspension and guaranteeing
successful transmission.

This CL also change the order to make sure the watchdog covers the whole
wakelock session, given that the binder call of requesting wakelock also
does context switching.

Bug: 422363726
Test: run_test.sh --host
Flag: EXEMPT, bugfix
Change-Id: I8d20a78d665fb5d6e890006b63f2721a7a193e82