camera vts: Fix Multi resolution image reader and reduced UHR sensor handling
CameraCharacteristics checks for Multi-resolution ImageReader didn't
take into account that its possible for a sensor to be a reduced ultra
high resolution sensor instead of a full ultra high resolution sensor.
With this CL we check for the presence of SENSOR_PIXEL_MODE is
availableRequestKeys to take that into account.

Bug: 424672169

Test: Vendor testing

Flag: EXEMPT; test fix

Change-Id: If2c81207b1806cd4555fd9ad108f1f74c3e93853
Signed-off-by: Jayant Chowdhary <jchowdhary@google.com>