Fix C++23 build: undefine NAN
math.h defines a NAN macro to __builtin_nanf(""). If this macro is
defined, then the IfaceType::NAN enumerator can't be used. The best
solution is to rename NAN (see NAN_IFACE, for example), but maybe we
can't do that for compatibility reasons, so the next best fix is to
add `#undef NAN` as necessary. There are other places in the Android
codebase that already do that.

Bug: 429048222
Test: m MODULES-IN-hardware-interfaces
Flag: EXEMPT bugfix
Change-Id: Ic74d4dbe74f70f5b275513758f0e081c45e5c8d7