audio: Resolve deadlock in the 'bluetooth' HAL module
There was a possibility for a deadlock to happen
when during execution of 'BluetoothAudioPortAidl::start'
the low latency support has not been established yet.
This is because 'getRecommendedLatencyModes' method also
takes the same lock.

The issue resolved by annotating with 'EXCLUDES' all
methods that take the 'mCvMutex' lock and ensuring that
no build errors are raised.

Flag: EXEMPT bugfix
Bug: 416413008
Test: atest VtsHalAudioCoreTargetTest
Change-Id: Icbc6a195303b0d02ada3002686167c8f94852d66