router_client: Rename HciRouterClientAgent registration methods
Rename `RegisterRouterClient` to `RegisterClient` and
`UnregisterRouterClient` to `UnregisterClient`.

The `Router` in the method names is redundant as the methods belong to
the `HciRouterClientAgent`. This change simplifies the method names for
better readability and conciseness.

Bug: 440204009
Test: sh run_test.sh
Flag: EXEMPT, simple refactor.
Updated all call sites and tests to reflect the new method names.

Change-Id: I3f71b58a0a89a18c0ab1da74492aaf87954f5ed9