Enforce @FixedSize for MQDescriptor type parameter
MQDescriptor is used with libfmq to set up a fast message queue. It
requires that its type paramter T be fixed size. Currently, this
requirement is not enforced by the AIDL compiler which leads to runtime
errors when a non-fixed sized type is used.

This commit applies the @FixedSize annotation to the type parameter T in
MQDescriptor and makes it fail immediately if a non-fixed sized type is
used.

Test: atest aidl_unittests
Bug: None
Test: EXEMPT build time host enforcement
Change-Id: Ie840500288e001fe227f126ad94932765318f61b