Separate file_path from id for mock video cameras
In the EVS Manager 1.1, users could provide a file path for static PNG
files for emulation, separate from the camera ID. With the move to the
AIDL version, the emulated video stream feature was relocated from EVS
Manager to the EVS default/Mock HAL, which now utilizes media files
instead of static PNGs. This change made the camera ID function as the
path to the video file, which is not ideal.

This change reintroduces the capability for users to specify a separate
`file_path` for video emulation in the XML configuration, distinct from
the camera ID. It adds the `file_path` attribute while retaining the
fallback to using the `id` as a path to the video, as this behaviour was
requested to be reimplemented.

Test: manual - changed camera configs on seahawk and checked camera
outputs and logs

Bug: 351856790
Change-Id: I8dcac5e3be321af58f1ccb7afafdc4f8aec73934