cs: Fallback to product-specific CS config
If the default channel sounding calibration config file is not found,
attempt to load a product-specific one.

The product-specific config file is named by appending the value of the
`ro.product.name` property to the default config file name. For example,
if the default file is `.../cs_config.json` and `ro.product.name` is
"product", the fallback file will be `.../cs_config_product.json`.

This allows for per-product customization of channel sounding
calibration parameters.

Bug: 434564115
Test: Manual.
Flag: EXEMPT, simple refactor.
Change-Id: I1145567d2b284b96bce2ff71088cdf3aed181df4