audio: Make Remote Submix VTS more robust
VTS tests for remote submix were relying implicitly on the
fact that audio profiles for input and output ports of remote
submix are exactly the same; shuffling the order of profiles
was breaking them.

Make the tests more robust by using the audio configuration
of already created stream for the second stream of the pair.

Bug: 434879819
Test: atest VtsHalAudioCoreTargetTest
Change-Id: Ibd9123335e60e9c49537a4eff15572658b01017b