audio: Add VTS opening remote submix with mismatching configs
Add tests that verify behavior of remote submix when the client
attempts to open the input and the output ends of the pipe with
mismatching configs. Two outcomes are allowed:

 - remote submix HAL may make it impossible to do this, for
   example by limiting the profile set of the other end to match
   the audio configuration used by the iniitally opened end;
 - or remote sumbix may support this case by adapting the audio
   data (for example, resampling, up/downmixing, etc), in which
   case the I/O over the remote submix pipe must work.

This is to avoid the regression happened in the past when
the default remote submix module implementation was allowing to
open ends with mismatching configs, but then failing only when
I/O was initiated.

Flag: EXEMPT TEST_ONLY
Bug: 434879819
Test: atest VtsHalAudioCoreTargetTest
Change-Id: Ibac13356931f6b6ba3a602b700ce4e0f56fdb07b